# backend/Web/wwwroot/swagger/specification.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Loppe Portalen API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/Authorization/Register": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "operationId": "Authorization_RegisterUser",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authorization/Login": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "operationId": "Authorization_AuthenticateUser",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateUserRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticateUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authorization/RefreshToken": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "operationId": "Authorization_RefreshToken",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshJwtRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshJwtResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Market/new": {
      "post": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_CreateMarket",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMarketRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMarketResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Market/instance/{id}": {
      "get": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_GetMarketInstance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMarketInstanceQueryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Market/instance/cancel/{id}": {
      "patch": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_CancelMarketInstance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelMarketInstanceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Market/instance/filtered": {
      "get": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_GetFilteredMarketInstances",
        "parameters": [
          {
            "name": "isCancelled",
            "in": "query",
            "schema": {
              "type": "boolean",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "organiserId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "categories",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string"
              }
            },
            "x-position": 5
          },
          {
            "name": "x",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "y",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "x-position": 7
          },
          {
            "name": "distance",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "x-position": 8
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFilteredMarketsQueryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Market/addStalls": {
      "post": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_AddStallsToMarket",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStallsToMarketRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddStallsToMarketResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Market/bookstalls": {
      "post": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_BookStalls",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookStallsRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookStallsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Market/banner/upload": {
      "put": {
        "tags": [
          "Market"
        ],
        "operationId": "Market_UploadMarketBanner",
        "parameters": [
          {
            "name": "marketId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadMarketBannerResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Merchant/booths": {
      "get": {
        "tags": [
          "Merchant"
        ],
        "operationId": "Merchant_GetBooths",
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoothBaseVM"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Merchant/add/contactInformation": {
      "post": {
        "tags": [
          "Merchant"
        ],
        "operationId": "Merchant_AddContactInformation",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMerchantContactInformationRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddContactInformationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Merchant/contactInfo": {
      "delete": {
        "tags": [
          "Merchant"
        ],
        "operationId": "Merchant_RemoveContactInformation",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveMerchantContactRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveMerchantContactResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Merchant/banner/upload": {
      "put": {
        "tags": [
          "Merchant"
        ],
        "operationId": "Merchant_UploadMerchantBanner",
        "parameters": [
          {
            "name": "merchantId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadMerchantBannerResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterUserResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AuthResult"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "id": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "AuthResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "token": {
                "type": "string",
                "nullable": true
              },
              "refreshToken": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
...
      "RegisterUserRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "country": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AuthenticateUserResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AuthResult"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "AuthenticateUserRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "RefreshJwtResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AuthResult"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "RefreshJwtRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "BoothBaseVM": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "stall": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/StallBaseVM"
              }
            ]
          },
          "categories": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateMarketResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "market": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CreateMarketVM"
              }
            ]
          }
        }
      },
      "CreateMarketVM": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MarketBaseVM"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "organiser": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/OrganiserBaseVM"
                  }
                ]
              }
            }
          }
        ]
      },
...
      "CreateMarketRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "organiserId": {
            "type": "integer",
            "format": "int32"
          },
          "marketName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "$ref": "#/components/schemas/Vector2"
          }
        }
      },
      "Vector2": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "x": {
            "type": "number",
            "format": "float"
          },
          "y": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "GetMarketInstanceQueryResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "market": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/GetMarketInstanceVM"
              }
            ]
          }
        }
      },
      "GetMarketInstanceVM": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MarketBaseVM"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "organiser": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/OrganiserBaseVM"
                  }
                ]
              },
              "stallTypes": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/StallTypeBaseVM"
                }
              },
              "booths": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/BoothBaseVM"
                }
              },
              "stalls": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/StallBaseVM"
                }
              },
              "imageData": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
...
      "CancelMarketInstanceResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "marketId": {
            "type": "integer",
            "format": "int32"
          },
          "isCancelled": {
            "type": "boolean"
          }
        }
      },
      "GetFilteredMarketsQueryResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "markets": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FilteredMarketVM"
            }
          }
        }
      },
      "FilteredMarketVM": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MarketBaseVM"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "organiser": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/OrganiserBaseVM"
                  }
                ]
              }
            }
          }
        ]
      },
...
      "AddStallsToMarketResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "stalls": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/StallBaseVM"
            }
          }
        }
      },
      "AddStallsToMarketRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "marketId": {
            "type": "integer",
            "format": "int32"
          },
          "stallTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BookStallsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "BookStallsRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "marketId": {
            "type": "integer",
            "format": "int32"
          },
          "merchantId": {
            "type": "integer",
            "format": "int32"
          },
          "stalls": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/StallBooking"
            }
          }
        }
      },
...
      "UploadMarketBannerResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "marketId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "AddContactInformationResponse": {
        "type": "object",
        "additionalProperties": false
      },
      "AddMerchantContactInformationRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "merchantId": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ContactInfoType"
          }
        }
      },
      "RemoveMerchantContactResponse": {
        "type": "object",
        "additionalProperties": false
      },
      "RemoveMerchantContactRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "merchantId": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UploadMerchantBannerResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "merchantId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}